Moved HttpServer to own namespace
