--- UltraStar Play/Assets/Common/Network/HttpServer/EndpointRequestData.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Net;
-
-public struct EndpointRequestData
-{
-    public Dictionary<string, string> PathParameters { get; private set; }
-    public HttpListenerContext Context { get; private set; }
-
-    public EndpointRequestData(HttpListenerContext context, Dictionary<string, string> pathParameters)
-    {
-        PathParameters = pathParameters;
-        Context = context;
-    }
-}
