+++ UltraStar Play/Assets/Editor/Tests/CurlyBracePlaceholderMatcherTest.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using NUnit.Framework;
+using SimpleHttpServerForUnity;
 
 public class CurlyBracePlaceholderMatcherTest
 {
