+++ UltraStar Play/Assets/Plugins/SimpleHttpServerForUnity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dc03b7226d2dd7b4ba175b25b685f942
+guid: fc5e8b025b539264faa2c6ac81cd4bff
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
