+++ UltraStar Play/Assets/Plugins/SimpleHttpServerForUnity/HttpServerExtensions.cs.meta
